Re-refactoring StopsHandler, using new lessons learned
